chore: delete Scripts/bootstrap.sh (#454)
BREAKING CHANGE: Scripts/bootstrap.sh is deleted as it is not needed anymore

This script confused me until I found out that it only writes the text
"Dependencies up-to-date" to the console. All the other code is only a
scaffold for code that was removed while removing Carthage in
1984360c64ab77ac621898c176b4a12fc4b03e02. Therefore I think that this
script is not needed anymore at all.